fix/ fix bugs, add training button
