+++ src/components/Library/LibraryComponent.styled.jsx
@@ -1,0 +1,19 @@
+import { Button } from 'antd';
+import styled from 'styled-components';
+
+export const StyledSection = styled.section`
+  padding: 40px 0;
+  font-family: 'Montserrat';
+`;
+
+export const StyledButton = styled(Button)`
+  display: block;
+  min-width: 200px;
+  min-height: 40px;
+  margin-left: auto;
+  margin-right: auto;
+
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 1.25;
+`;
