+++ src/components/SharedLayout/SharedLayout.jsx
@@ -1,17 +1,17 @@
 import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 import Header from 'components/Header';
-import Container from 'components/Container';
+// import Container from 'components/Container';
 import Loader from 'components/Loader';
 
 const SharedLayout = () => {
   return (
-    <Container>
+    <>
       <Header />
       <Suspense fallback={<Loader />}>
         <Outlet />
       </Suspense>
-    </Container>
+    </>
   );
 };
